requirements.txt
google-generativeai
yt-dlp
groq
Flask
python-dotenv